+++ credo-bank.Application/Models/DTO/Messages/LoanApplicationSubmitted.cs
@@ -6,7 +6,7 @@ public class LoanApplicationSubmitted
 {
     public int LoanId { get; set; }
     public int UserId { get; set; }
-    public int LoanAmount { get; set; }
+    public decimal LoanAmount { get; set; }
     public int LoanTermInMonths { get; set; }
     public Currency CurrencyType { get; set; }
     public Domain.Enums.Application ApplicationStatus { get; set; }
