+++ credo-bank.Application/Models/DTO/Response/LoanApplicationDto.cs
@@ -5,7 +5,7 @@ namespace credo_bank.Application.Models.DTO.Response;
 public class LoanApplicationDto
 {
     public int Id { get; set; }
-    public int LoanAmount { get; set; }
+    public decimal LoanAmount { get; set; }
     public int LoanTermInMonths { get; set; }
     public Currency CurrencyType { get; set; }
     public Domain.Enums.Application ApplicationStatus { get; set; }
